Move pagination components to `core` folder
