+++ src/components/core/grid.js
@@ -3,12 +3,13 @@ import styled from 'styled-components'
 const gutter = '1rem'
 
 export const Grid = styled.div`
+  display: flex;
+  flex-wrap: wrap;
   margin-top: -${gutter};
   margin-left: -${gutter};
 `
 
 export const Cell = styled.div`
-  display: inline-block;
   padding-top: ${gutter};
   padding-left: ${gutter};
 `
