+++ src/components/core/pagination/index.js
