+++ src/components/core/pagination/pagination.test.js
